+++ NSplatformer/GameView.h
@@ -7,13 +7,13 @@ class GameView
 public:
 	void setPos(const sf::Vector2f& pos);
 	void setSize(const sf::Vector2f& size);
-	void setWorldSize(const sf::Vector2f& size);
+	void setWorldSize(const sf::Vector2i& size);
 	void moveCamera(const sf::Vector2f& pos, const sf::Vector2f& deltaPos);
 	void setView(sf::RenderWindow& window);
 	bool isInView(const sf::Vector2f& pos, const sf::Vector2f& size) const;
 private:
 	sf::View m_view;
 	sf::Vector2f m_size;
-	sf::Vector2f m_worldSize;
+	sf::Vector2i m_worldSize;
 };
